Fix #8022: Refactor MoveStatics
We still keep two phases, but the code is simplified.

We cannot merge MoveStatics with CheckStatic because the early phases
are not able to handle static calls, thus it has to be at the end of
the pipeline.  Meanwhile, we want checks to happen as early as
possible, thus CheckStatic has to be in the beginning of the pipeline.
