+++ compiler/src/dotty/tools/dotc/transform/MoveStatics.scala
@@ -1,26 +1,29 @@
-package dotty.tools.dotc.transform
+package dotty.tools.dotc
+package transform
 
-import dotty.tools.dotc.ast.{Trees, tpd}
-import dotty.tools.dotc.core.Annotations.Annotation
-import dotty.tools.dotc.core.Contexts._
-import dotty.tools.dotc.core.DenotTransformers.SymTransformer
-import dotty.tools.dotc.core.SymDenotations.SymDenotation
-import dotty.tools.dotc.core.NameOps._
-import dotty.tools.dotc.core.Flags
-import dotty.tools.dotc.core.Names.Name
-import dotty.tools.dotc.core.StdNames.nme
-import dotty.tools.dotc.core.Symbols._
-import dotty.tools.dotc.core.Types.MethodType
-import dotty.tools.dotc.transform.MegaPhase.MiniPhase
+import core._
+import Flags._
+import Contexts._
+import Symbols._
+import Decorators._
+import DenotTransformers.SymTransformer
+import Types.MethodType
+import Annotations.Annotation
+import SymDenotations.SymDenotation
+import Names.Name
+import StdNames.nme
+import NameOps._
 
-object MoveStatics {
-  val name: String = "moveStatic"
-}
+import reporting._
+import ast._
+
+import SymUtils._
+import MegaPhase._
 
 /** Move static methods from companion to the class itself */
 class MoveStatics extends MiniPhase with SymTransformer {
+  import ast.tpd._
 
-  import tpd._
   override def phaseName: String = MoveStatics.name
 
   def transformSym(sym: SymDenotation)(using Context): SymDenotation =
@@ -32,14 +35,30 @@ class MoveStatics extends MiniPhase with SymTransformer {
     }
     else sym
 
+  override def transformSelect(tree: tpd.Select)(using Context): tpd.Tree =
+    if (tree.symbol.hasAnnotation(defn.ScalaStaticAnnot)) {
+      def isSafeQual(t: Tree): Boolean = // follow the desugared paths created by typer
+        t match {
+          case t: This => true
+          case t: Select => isSafeQual(t.qualifier)
+          case t: Block => t.stats.forall(tpd.isPureExpr) && isSafeQual(t.expr)
+          case _ => false
+        }
+
+      if (isSafeQual(tree.qualifier))
+        ref(tree.symbol)
+      else
+        Block(tree.qualifier :: Nil, ref(tree.symbol))
+    }
+    else tree
+
+
   override def transformStats(trees: List[Tree])(using Context): List[Tree] =
     if (ctx.owner.is(Flags.Package)) {
       val (classes, others) = trees.partition(x => x.isInstanceOf[TypeDef] && x.symbol.isClass)
       val pairs = classes.groupBy(_.symbol.name.stripModuleClassSuffix).asInstanceOf[Map[Name, List[TypeDef]]]
 
       def rebuild(orig: TypeDef, newBody: List[Tree]): Tree = {
-        if (orig eq null) return EmptyTree
-
         val staticFields = newBody.filter(x => x.isInstanceOf[ValDef] && x.symbol.hasAnnotation(defn.ScalaStaticAnnot)).asInstanceOf[List[ValDef]]
         val newBodyWithStaticConstr =
           if (staticFields.nonEmpty) {
@@ -61,21 +80,30 @@ class MoveStatics extends MiniPhase with SymTransformer {
         assert(companion != module)
         if (!module.symbol.is(Flags.Module)) move(companion, module)
         else {
-          val allMembers =
-            (if (companion != null) {companion.rhs.asInstanceOf[Template].body} else Nil) ++
-            module.rhs.asInstanceOf[Template].body
-          val (newModuleBody, newCompanionBody) = allMembers.partition(x => {assert(x.symbol.exists); x.symbol.owner == module.symbol})
-          Trees.flatten(rebuild(companion, newCompanionBody) :: rebuild(module, newModuleBody) :: Nil)
+          val moduleTmpl = module.rhs.asInstanceOf[Template]
+          val companionTmpl = companion.rhs.asInstanceOf[Template]
+          val (staticDefs, remainingDefs) = moduleTmpl.body.partition {
+            case memberDef: MemberDef => memberDef.symbol.isScalaStatic
+            case _ => false
+          }
+
+          rebuild(companion, companionTmpl.body ++ staticDefs) :: rebuild(module, remainingDefs) :: Nil
         }
       }
       val newPairs =
         for ((name, classes) <- pairs)
           yield
-            if (classes.tail.isEmpty)
-              if (classes.head.symbol.is(Flags.Module)) move(classes.head, null)
-              else List(rebuild(classes.head, classes.head.rhs.asInstanceOf[Template].body))
+            if (classes.tail.isEmpty) {
+              val classDef = classes.head
+              val tmpl = classDef.rhs.asInstanceOf[Template]
+              rebuild(classDef, tmpl.body) :: Nil
+            }
             else move(classes.head, classes.tail.head)
       Trees.flatten(newPairs.toList.flatten ++ others)
     }
     else trees
+}
+
+object MoveStatics {
+  val name: String = "moveStatic"
 }
