+++ tests/run/statics.scala
@@ -1,30 +1,29 @@
 import scala.annotation.static
 
-class Foo{
+class Foo {
   class Bar {
     def qwa =
-     Bar.field
+      Bar.field
          // 0: invokestatic  #31                 // Method Foo$Bar$.field:()I
          // 3: ireturn
   }
   object Bar {
-     @static
      val field = 1
    }
 }
 
-object Foo{
- @static
- def method = 1
+object Foo {
+  @static
+  def method = 1
 
- @static
- val field = 2
+  @static
+  val field = 2
 
- @static
- var mutable = 3
+  @static
+  var mutable = 3
 
- @static
- def accessor = field
+  @static
+  def accessor = field
 }
 
 object Test {
@@ -34,7 +33,7 @@ object Test {
  }
 }
 
-class WithLazies{
+class WithLazies {
   lazy val s = 1
   //         98: getstatic     #30                 // Field WithLazies$.OFFSET$0:J
 }
